+++ CHANGELOG.md
@@ -6,6 +6,49 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Breaking Change: Remove `SelectWithState` (replaced by `DrodownSelect`)
 
+## [3.0.0-beta.4] - 2021-02-22
+
+Breaking changes:
+
+- Delete the following deprecated SCSS mixins:
+  - `k-Badge`
+  - `k-Header`
+  - `k-IconBadge`
+  - `k-ImageWithCaption`
+  - `k-InformationBox`
+  - `k-PhoneDropdown`
+  - `k-Separator`
+  - `k-separatorBase`
+  - `k-VerticalSeparator`
+  - `k-VerticalStroke`
+- Delete the following deprecated components:
+  - `Comment` from `form/comment`
+    (replaced by `comments/comment`)
+  - `CommentAvatar` from `form/comment-avatar`
+    (replaced by `comments/comment-avatar`)
+  - `CommentForm` from `form/comment-form`
+    (replaced by `comments/comment-form`)
+  - `Dropdown`
+  - `DropdownButton`
+  - `Header` (replaced by `HeaderNav`)
+  - `HeaderLogo` (replaced by `HeaderNav`)
+  - `HeaderTitles` (replaced by `HeaderNav`)
+  - `ImageWithCaption`
+  - `NavList`
+  - `PhoneDropdown`
+  - `Switch` (replaced by `ToggleSwitch`)
+  - `UserMenu`
+- Convert and move the following components to CSS-in-JS:
+  - `Badge`
+  - `IconBadge`
+
+Features:
+
+- Convert the following components to CSS-in-JS:
+  - `InformationBox`
+  - `VerticalStroke`
+- Add `onChange` to `ToggleSwitch` API.
+
 ## [3.0.0-beta.3] - 2021-02-19
 
 Breaking changes:
@@ -88,6 +131,20 @@ Breaking changes:
   - `ProjectSimilarCard`
   - `RewardCardLegacy`
   - `SimilarProjectsCard`.
+
+## [2.121.0] - 2021-02-22
+
+Features:
+
+- Add `DashboardLayout` component with `Flow` subcomponent.
+- Add `DashboardMenu` component.
+- Add `LightbulbIllustration` with `size` (large, normal) prop.
+- Add icons components:
+  - `HomeIcon`
+  - `TagIcon`
+  - `FileIcon`
+  - `LoudspeakerIcon`
+  - `ShieldIcon`.
 
 ## [2.120.1] - 2021-02-19